chore: UIColor+Hex를 UIColor+로  통합
