+++ Koin/Core/Extensions/UIKit/UIColor+.swift
@@ -27,3 +27,24 @@ extension UIColor {
         return UIColor(hue: hue, saturation: saturation, brightness: brightness, alpha: 1.0)
     }
 }
+
+extension UIColor {
+    convenience init(hexCode: String, alpha: CGFloat = 1.0) {
+        let hexCode = hexCode.trimmingCharacters(in: .whitespacesAndNewlines)
+        let scanner = Scanner(string: hexCode)
+        
+        if hexCode.hasPrefix("#") {
+            scanner.currentIndex = hexCode.index(after: hexCode.startIndex)
+        }
+        
+        var rgb: UInt64 = 0
+        scanner.scanHexInt64(&rgb)
+        
+        let red = CGFloat((rgb >> 16) & 0xFF) / 255.0
+        let green = CGFloat((rgb >> 8) & 0xFF) / 255.0
+        let blue = CGFloat(rgb & 0xFF) / 255.0
+        
+        self.init(red: red, green: green, blue: blue, alpha: alpha)
+    }
+}
+
